chore: 升级 Gradle 至 9.3.1 并更新相关依赖
- 升级 Gradle Wrapper 至 9.3.1，Android Gradle Plugin 至 9.0.0
- 升级 Kotlin 至 2.3.20，并配置 JVM Toolchain 17
- 更新 BaseRecyclerViewAdapterHelper 至 v4，适配新的 ViewHolder API
- 将 JSON 序列化从 Gson 迁移至 kotlinx.serialization
- 添加 Navigation 3 和 Compose 相关依赖
- 修复因 Kotlin 2.x 与 Java 17 导致的 R8 编译问题
- 优化构建脚本，移除废弃配置项

+++ app/src/main/java/shetj/me/base/view/RadarView.kt
@@ -117,14 +117,14 @@ class RadarView @JvmOverloads constructor(
 
         for (i in 0 until circleCount) {
             val circleRadius = radius - i * circleSpacing
-            canvas.drawCircle(centerX.toFloat(), centerY, circleRadius.toFloat(), circlePaint)
+            canvas.drawCircle(centerX, centerY, circleRadius, circlePaint)
         }
 
 
         // 画从圆心到每个顶点的线条
         paint.color = ContextCompat.getColor(context, R.color.c_2)
         points.forEachIndexed { index, point ->
-            canvas.drawLine(centerX.toFloat(), centerY.toFloat(), point.x, point.y, paint)
+            canvas.drawLine(centerX, centerY, point.x, point.y, paint)
         }
